+++ NuclearSource.h
@@ -2,9 +2,14 @@
 #define NUCLEARSOURCE_H
 
 #include "EnergySource.h"
+/**
+ * @file NuclearSource.h
+ * 
+ * @brief The header file for the NuclearSource class
+ */
 class NuclearSource : public EnergySource {
 public:
-	NuclearSource(int output);
+	NuclearSource(float output);
 
 	float getEnergyOutput();
